requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
logging==0.4.9.6
openai
